refactor: remove config module and integrate settings directly into server
