--- config/config.js
@@ -1,84 +1,0 @@
-/*
- * Config file: This file no longer requires manual updates to values,
- * as they are now managed through environment variables.
- */
-import dotenv from "dotenv";
-
-// Load environment variables from .env file
-dotenv.config({ quiet: true });
-
-const getBackendConfig = async function (email) {
-  /*
-    This function should:
-    - Accept a user or customer identifier (such as email)
-    - Look up the correct Qlik Cloud tenant for that customer
-    - Retrieve the corresponding OAuth client details for the backend activity (review guiding
-      principles for OAuth M2M impersonation: https://qlik.dev/authenticate/oauth/guiding-principles-oauth-impersonation/)
-  
-    For purposes of making this demo as simple as possible, the values are hardcoded
-    in this project via a .env file. Do not do this in production.
-
-    */
-
-  // Build app settings
-  const appSettings = {
-    secret: process.env.SESSION_SECRET,
-    port: process.env.PORT,
-    userPrefix: process.env.USER_PREFIX,
-    hypercubeDimension: process.env.HYPERCUBE_DIMENSION,
-    hypercubeMeasure: process.env.HYPERCUBE_MEASURE,
-  };
-
-  // Build qlik/api backend config
-  const configBackend = {
-    authType: "oauth2",
-    host: process.env.TENANT_URI,
-    clientId: process.env.OAUTH_BACKEND_CLIENT_ID,
-    clientSecret: process.env.OAUTH_BACKEND_CLIENT_SECRET,
-    noCache: true,
-  };
-
-  // Build qlik/api frontend config
-  const configFrontend = {
-    authType: "oauth2",
-    host: process.env.TENANT_URI,
-    clientId: process.env.OAUTH_FRONTEND_CLIENT_ID,
-    clientSecret: process.env.OAUTH_FRONTEND_CLIENT_SECRET,
-    noCache: true,
-  };
-
-  return { configBackend, configFrontend, appSettings };
-};
-
-const getFrontendConfig = async function (email) {
-  /*
-    This function should:
-    - Accept a user or customer identifier (such as email)
-    - Look up the correct Qlik Cloud tenant for that customer
-    - Retrieve the corresponding impersonation OAuth client for the qlik-embed tag (review guiding
-      principles for OAuth M2M impersonation: https://qlik.dev/authenticate/oauth/guiding-principles-oauth-impersonation/)
-  
-    For purposes of making this demo as simple as possible, the values are hardcoded
-    in this project via a .env file. Do not do this in production.
-
-    */
-
-  // This is a cut-down config without the client secret for use in the qlik-embed HEAD tag.
-  const myParamsConfig = {
-    tenantUri: process.env.TENANT_URI,
-    oAuthFrontEndClientId: process.env.OAUTH_FRONTEND_CLIENT_ID,
-    appId: process.env.APP_ID,
-    sheetId: process.env.SHEET_ID,
-    objectId: process.env.OBJECT_ID,
-    fieldId: process.env.FIELD_ID,
-    assistantId: process.env.ASSISTANT_ID,
-    hypercubeDimension: process.env.HYPERCUBE_DIMENSION,
-    hypercubeMeasure: process.env.HYPERCUBE_MEASURE,
-    masterDimension: process.env.MASTER_DIMENSION,
-    masterMeasure: process.env.MASTER_MEASURE
-  };
-
-  return { myParamsConfig };
-};
-
-export { getFrontendConfig, getBackendConfig };
